ABI_V0 compatibility: graphics.library function offsets and protos

+++ rom/graphics/draw.c
@@ -57,8 +57,8 @@ static ULONG draw_render(APTR draw_rd, WORD srcx, WORD srcy,
 
 /*  SYNOPSIS */
 	AROS_LHA(struct RastPort *, rp, A1),
-	AROS_LHA(WORD             , x, D0),
-	AROS_LHA(WORD             , y, D1),
+	AROS_LHA(LONG             , x, D0),
+	AROS_LHA(LONG             , y, D1),
 
 /*  LOCATION */
 	struct GfxBase *, GfxBase, 41, Graphics)
