+++ rom/graphics/loadrgb4.c
@@ -18,7 +18,7 @@
 /*  SYNOPSIS */
 	AROS_LHA(struct ViewPort *, vp, A0),
 	AROS_LHA(UWORD           *, colors, A1),
-	AROS_LHA(WORD             , count, D0),
+	AROS_LHA(LONG             , count, D0),
 
 /*  LOCATION */
 	struct GfxBase *, GfxBase, 32, Graphics)
